Move gas_cost mod to a separate file
